requirements.txt
Click==7.0
dnspython==1.16.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.0
MarkupSafe==1.1.1
numpy==1.17.4
pandas==0.25.3
pymongo==3.9.0
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
scikit-learn==0.22
scipy==1.3.3
six==1.13.0
sklearn==0.0
Werkzeug==0.16.0
